pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "djauthij"
version = "0.1.0"
description = "Reusable Django authentication app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Django>=4.2",
    "django-huey>=1.3.1"
]

[tool.setuptools.packages.find]
where = ["."]